refactor(api-gateway): Phase 8 carve BatchReplayFailures + delete OperatorModule
- New Features/Operator/BatchReplayFailures/{Endpoint,Handler,
  BatchReplayRequest,BatchReplayItem,BatchReplayResponse,
  BatchReplayFailuresSliceExtensions} registers POST /operator/api/failures/replay-batch
  via per-slice extensions; handler resolves replayedBy with same JwtClaimTypes
  fallback chain and preserves byte-identical per-item status mapping
  (success|not_found|not_pending|publish_failed|unknown).
- Program.cs now owns the shared DI (AddPlatformEventBus, AddDeadLetter,
  AddRequireOperatorPolicy) and OutboxPollerOptions binding + conditional
  poller registration that previously lived in OperatorModule.AddServices,
  then chains AddBatchReplayFailuresSlice + MapBatchReplayFailuresSlice.
- Deleted ApiGateway/Operator/OperatorModule.cs and the now-empty Operator/
  folder; GatewayTestHarness swaps OperatorModule.MapEndpoints for the new
  slice mapping; DeadLetterReplayBatchEndpointTests now drives the new
  BatchReplayFailuresHandler through a ClaimsPrincipal helper;
  OperatorMessagingProviderTests inlines the same shared DI calls instead
  of going through the deleted module.
