+++ api-gateway/ApiGateway/Features/Operator/BatchReplayFailures/BatchReplayFailuresSliceExtensions.cs
@@ -1,0 +1,10 @@
+namespace ApiGateway.Features.Operator.BatchReplayFailures;
+
+internal static class BatchReplayFailuresSliceExtensions
+{
+    public static IServiceCollection AddBatchReplayFailuresSlice(this IServiceCollection services)
+    {
+        services.AddScoped<BatchReplayFailuresHandler>();
+        return services;
+    }
+}
