+++ api-gateway/ApiGateway/Features/Operator/BatchReplayFailures/BatchReplayItem.cs
@@ -1,0 +1,3 @@
+namespace ApiGateway.Features.Operator.BatchReplayFailures;
+
+public sealed record BatchReplayItem(Guid Id, string Status, Guid? NewMessageId, string? Reason);
