+++ api-gateway/ApiGateway/Features/Operator/BatchReplayFailures/BatchReplayRequest.cs
@@ -1,0 +1,3 @@
+namespace ApiGateway.Features.Operator.BatchReplayFailures;
+
+public sealed record BatchReplayRequest(IReadOnlyList<Guid> Ids);
