+++ api-gateway/ApiGateway/Features/Operator/BatchReplayFailures/BatchReplayResponse.cs
@@ -1,0 +1,3 @@
+namespace ApiGateway.Features.Operator.BatchReplayFailures;
+
+public sealed record BatchReplayResponse(IReadOnlyList<BatchReplayItem> Items);
